+++ test/audits/forms/oppty-handlers/low-views-oppoty-handler.test.js
@@ -14,10 +14,10 @@
 import { expect, use } from 'chai';
 import sinon from 'sinon';
 import sinonChai from 'sinon-chai';
-import createLowViewsOpportunities from '../../../src/forms-opportunities/oppty-handlers/low-views-handler.js';
-import { FORM_OPPORTUNITY_TYPES } from '../../../src/forms-opportunities/constants.js';
-import testData from '../../fixtures/forms/high-form-views-low-conversions.js';
-import { DATA_SOURCES } from '../../../src/common/constants.js';
+import createLowViewsOpportunities from '../../../../src/forms-opportunities/oppty-handlers/low-views-handler.js';
+import { FORM_OPPORTUNITY_TYPES } from '../../../../src/forms-opportunities/constants.js';
+import testData from '../../../fixtures/forms/high-form-views-low-conversions.js';
+import { DATA_SOURCES } from '../../../../src/common/constants.js';
 
 use(sinonChai);
 describe('createLowFormViewsOpportunities handler method', () => {
@@ -79,10 +79,10 @@ describe('createLowFormViewsOpportunities handler method', () => {
       runbook: 'https://adobe.sharepoint.com/:w:/s/AEM_Forms/EeYKNa4HQkRAleWXjC5YZbMBMhveB08F1yTTUQSrP97Eow?e=cZdsnA',
       type: FORM_OPPORTUNITY_TYPES.LOW_VIEWS,
       origin: 'AUTOMATION',
-      title: 'The form has low views',
+      title: 'Form has low views',
       description: 'The form has low views but the page containing the form has higher traffic',
       tags: [
-        'Forms Conversion',
+        'Form View',
       ],
       data: {
         form: 'https://www.surest.com/high-page-low-form-view',
@@ -136,7 +136,7 @@ describe('createLowFormViewsOpportunities handler method', () => {
       guidance: {
         recommendations: [
           {
-            insight: 'The form in the page: https://www.surest.com/high-page-low-form-view has low discoverability and only 2.9895366218236172% visitors landing on the page are viewing the form.',
+            insight: 'The form in the page: https://www.surest.com/high-page-low-form-view has low discoverability and only 2.99% visitors landing on the page are viewing the form.',
             recommendation: 'Position the form higher up on the page so users see it without scrolling. Consider using clear and compelling CTAs, minimizing distractions, and ensuring strong visibility across devices.',
             type: 'guidance',
             rationale: 'Forms that are visible above the fold are more likely to be seen and interacted with by users.',
@@ -167,7 +167,7 @@ describe('createLowFormViewsOpportunities handler method', () => {
       {
         recommendations: [
           {
-            insight: 'The form in the page: https://www.surest.com/existing-opportunity has low discoverability and only 2.9895366218236172% visitors landing on the page are viewing the form.',
+            insight: 'The form in the page: https://www.surest.com/existing-opportunity has low discoverability and only 2.99% visitors landing on the page are viewing the form.',
             recommendation: 'Position the form higher up on the page so users see it without scrolling. Consider using clear and compelling CTAs, minimizing distractions, and ensuring strong visibility across devices.',
             type: 'guidance',
             rationale: 'Forms that are visible above the fold are more likely to be seen and interacted with by users.',